// src/catalog/aws/aurora.gen.ts
type Rates = readonly [number | null, number | null, number | null, number | null, number | null, number | null]

export interface AuroraPriceRow {
  readonly vcpu: number
  readonly memoryGiB: number
  readonly od: number
  readonly ioOptimized: number | null
  readonly ri: Rates | null
}

export const auroraPrices = {
  postgresql: {
// ...
    'db.r7g.large': { vcpu: 2, memoryGiB: 16, od: 0.276, ioOptimized: 0.359, ri: [0.2128, 0.207253, 0.201712, null, 0.151989, 0.146499] },
// ...
  },
// ...
} as const satisfies Record<string, Record<string, AuroraPriceRow>>

export const auroraStorageRates = {
  postgresql: {
    io: 2e-7,
    ioOptimizedStorage: 0.225,
    storage: 0.1,
  },
  mysql: {
    io: 2e-7,
    ioOptimizedStorage: 0.225,
    storage: 0.1,
  },
} as const